+++ HandlerOnePort.h
@@ -12,7 +12,7 @@ class ClientHandlerOnePort;
 class HandlerOnePort : public IHandler {
 	map<string, ClientHandlerOnePort*> clients;
 protected:
-	void acceptNewClient(int, char*, int, sockaddr_in*, socklen_t*);
+	void acceptNewClient(int, char*, ssize_t, sockaddr_in*, socklen_t*);
 public:
 	HandlerOnePort(IOManager * io) : IHandler(io) {}
 	void handle(int, bool);
